refactor(google-setup): extract shared helpers to shared.mjs
Move ask(), runCommand(), runCommandOutput(), and GCP_APIS out of setup.mjs
into a new shared.mjs so both setup.mjs and the upcoming admin-setup.mjs
can import them without duplication.
