+++ services/kiloclaw/google-setup/setup.mjs
@@ -15,9 +15,9 @@
  *   docker run -it ghcr.io/kilo-org/google-setup --token=<jwt>
  */
 
-import { spawn, execSync, execFileSync } from 'node:child_process';
+import { execSync, execFileSync } from 'node:child_process';
 import crypto from 'node:crypto';
-import readline from 'node:readline';
+import { ask, runCommand, runCommandOutput, GCP_APIS } from './shared.mjs';
 
 // ---------------------------------------------------------------------------
 // CLI args
@@ -74,54 +74,6 @@ const authHeaders = {
 function workerApiUrl(path) {
   const base = `${workerUrl}${path}`;
   return instanceId ? `${base}?instanceId=${encodeURIComponent(instanceId)}` : base;
-}
-
-// APIs to enable in the GCP project
-const GCP_APIS = [
-  'gmail.googleapis.com',
-  'calendar-json.googleapis.com',
-  'drive.googleapis.com',
-  'docs.googleapis.com',
-  'slides.googleapis.com',
-  'sheets.googleapis.com',
-  'tasks.googleapis.com',
-  'people.googleapis.com',
-  'forms.googleapis.com',
-  'chat.googleapis.com',
-  'classroom.googleapis.com',
-  'script.googleapis.com',
-  'pubsub.googleapis.com',
-];
-
-// ---------------------------------------------------------------------------
-// Helpers
-// ---------------------------------------------------------------------------
-
-function ask(question) {
-  const rl = readline.createInterface({ input: process.stdin, output: process.stdout });
-  return new Promise(resolve => {
-    rl.question(question, answer => {
-      rl.close();
-      resolve(answer.trim());
-    });
-  });
-}
-
-function runCommand(cmd, args, opts = {}) {
-  return new Promise((resolve, reject) => {
-    const child = spawn(cmd, args, { stdio: 'inherit', ...opts });
-    child.on('close', code =>
-      code === 0 ? resolve() : reject(new Error(`${cmd} exited with code ${code}`))
-    );
-    child.on('error', reject);
-  });
-}
-
-function runCommandOutput(cmd, args) {
-  return execFileSync(cmd, args, {
-    encoding: 'utf8',
-    stdio: ['pipe', 'pipe', 'pipe'],
-  }).trim();
 }
 
 // ---------------------------------------------------------------------------
